+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.4 [24/02/2025]
+- Show if activity, service or/and provider is exported or not
+- Overwrite existing extraction folder
+- Fixed bugs
+- Improved code
+
 ## v1.3 [25/01/2025]
 - Changed color scheme
 - Changed banner